some int casts to fix a bug from sourceforge. see here for details:
http://sourceforge.net/tracker/index.php?func=detail&aid=522990&group_id=5113&atid=105113

+++ web/week.php
@@ -189,7 +189,9 @@ else for ($i = 0; ($row = sql_row($res, $i)); $i++)
 	# Note: int casts on database rows for min and max is needed for PHP3.
 
 	$t = max(round_t_down($row[0], $resolution), $week_start);
-	$end_t = min(round_t_up($row[1], $resolution), $week_end+1);
+	$end_t = min((int)round_t_up((int)$row[1],
+				     (int)$resolution), 
+		                     (int)$week_end+1);
 	$weekday = (date("w", $t) + 7 - $weekstarts) % 7;
 	$prev_weekday = -1; # Invalid value to force initial label.
 	$slot = ($t - $week_midnight) % 86400 / $resolution;
